requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-tavily
langchain-chroma
tavily-python
langgraph
langsmith
chromadb
pypdf
python-dotenv
pandas
kaggle
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openfoodfacts
